+++ modules/core/04-channel/keeper/grpc_query.go
@@ -401,7 +401,7 @@ func (q Keeper) UnreceivedPackets(c context.Context, req *types.QueryUnreceivedP
 	if !found {
 		return nil, status.Error(
 			codes.NotFound,
-			errorsmod.Wrapf(types.ErrChannelNotFound, "port-id: %s, channel-id %s", req.PortId, req.ChannelId).Error(),
+			sdkerrors.Wrapf(types.ErrChannelNotFound, "port-id: %s, channel-id %s", req.PortId, req.ChannelId).Error(),
 		)
 	}
 
@@ -424,7 +424,7 @@ func (q Keeper) UnreceivedPackets(c context.Context, req *types.QueryUnreceivedP
 		if !found {
 			return nil, status.Error(
 				codes.NotFound,
-				errorsmod.Wrapf(
+				sdkerrors.Wrapf(
 					types.ErrSequenceReceiveNotFound,
 					"destination port: %s, destination channel: %s", req.PortId, req.ChannelId,
 				).Error(),
@@ -445,7 +445,7 @@ func (q Keeper) UnreceivedPackets(c context.Context, req *types.QueryUnreceivedP
 	default:
 		return nil, status.Error(
 			codes.InvalidArgument,
-			errorsmod.Wrapf(types.ErrInvalidChannelOrdering, "channel order %s is not supported", channel.Ordering.String()).Error())
+			sdkerrors.Wrapf(types.ErrInvalidChannelOrdering, "channel order %s is not supported", channel.Ordering.String()).Error())
 	}
 
 	selfHeight := clienttypes.GetSelfHeight(ctx)